cob_base: added PLatformPArams class and Param_Loader for base_drive_chain
