+++ cob_base_drive_chain/ros/src/cob_base_drive_chain.cpp
@@ -77,6 +77,7 @@
 
 // external includes
 #include <cob_base_drive_chain/CanCtrlPltfCOb3.h>
+#include <cob_base_drive_chain/cob_base_drive_chain_param_loader.h>
 #include <cob_utilities/IniFile.h>
 #include <cob_utilities/MathSup.h>
